Split up the root reducer for cards and columns
